+++ .gitignore
@@ -2,3 +2,4 @@
 /dist/
 *.egg-info/
 tags
+.coverage
